moved summary tab related assets to their own folder
