+++ js/templates/modals/orderDetail/summaryTab/stateProgressBar.html
