+++ js/templates/modals/orderDetail/summaryTab/summary.html
@@ -10,4 +10,5 @@
 <div class="js-statusProgressBarContainer statusProgressBarContainer"></div>
 <hr class="clrBr rowLg" />
 
-<div class="js-paymentWrap"></div>
+<div class="js-paymentWrap row"></div>
+<div class="js-paymentWrap2"></div>
