+++ js/views/modals/orderDetail/summaryTab/Summary.js
@@ -1,9 +1,9 @@
-import app from '../../../app';
+import app from '../../../../app';
 import { clipboard } from 'electron';
-import '../../../utils/velocity';
-import loadTemplate from '../../../utils/loadTemplate';
+import '../../../../utils/velocity';
+import loadTemplate from '../../../../utils/loadTemplate';
 import { Model } from 'backbone';
-import BaseVw from '../../baseVw';
+import BaseVw from '../../../baseVw';
 import StateProgressBar from './StateProgressBar';
 import Payment from './Payment';
 
@@ -54,7 +54,7 @@ export default class extends BaseVw {
   }
 
   render() {
-    loadTemplate('modals/orderDetail/summary.html', t => {
+    loadTemplate('modals/orderDetail/summaryTab/summary.html', t => {
       this.$el.html(t({
         id: this.model.id,
         ...this.model.toJSON(),
@@ -79,7 +79,7 @@ export default class extends BaseVw {
           confirmations: 3537,
         }),
         initialState: {
-          paymentNumber: 1,
+          paymentNumber: 2,
           amountShort: 0,
           showAmountShort: false,
           payee: app.profile.get('name'),
@@ -87,6 +87,23 @@ export default class extends BaseVw {
         },
       });
       this.$('.js-paymentWrap').html(this.payment.render().el);
+
+      if (this.payment2) this.payment2.remove();
+      this.payment2 = this.createChild(Payment, {
+        model: new Model({
+          txid: '6d2cf390834a5578fdfe2bd2d2469992cce7d7c6656122ff78b968f62e2c41a4',
+          value: 0.00583,
+          confirmations: 4967856,
+        }),
+        initialState: {
+          paymentNumber: 1,
+          amountShort: 0.00032,
+          showAmountShort: true,
+          payee: app.profile.get('name'),
+          showActionButtons: true,
+        },
+      });
+      this.$('.js-paymentWrap2').html(this.payment2.render().el);
     });
 
     return this;
